+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+Nothing yet.
+
+## [v1.1.0] - 2018-08-18
+
 ### Added
 
 - `Player::events(&self)` returns a blocking iterator of player events.
@@ -78,3 +82,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased]: https://github.com/Mange/mpris-rs/compare/v1.0.0...HEAD
 [v1.0.0]: https://github.com/Mange/mpris-rs/compare/v0.1.0...v1.0.0
+[v1.1.0]: https://github.com/Mange/mpris-rs/compare/v1.0.0...v1.1.0